Arrumado coisa pra krl mas vai ai um resumo:
Os tipos das entitys estava errado (Long) então mudei todo os id para Integer, além das referências de Long para os métodos.

O Controller estava com erro no mapping, então defini um mapping para cada funçao de forma explicita, fica o mapping no inicio da classe + o mapping da função. Exemplo -> /Clientes/Listar

Descobri que para criar selects personalizados pelo JPA precisa estar com exatamente o mesmo nome do atributo, então acertei os que estavam errados.

Alterei o banco de dados para ficar mais fácil de mexer pelo Java. Troquei sexo bit por sexo boolean e aumentei o varchar do numero_tel.

Caso o service de cliente recebesse email null, ele dava erro pq não tinha tratamento para isso, então ele tentava usar algumas funções mesmo null, também arrumei.

Testei as alterações pelo postman e consegui dar insert.

FIM UFA

+++ backend/src/main/java/com/ib/alocacao_carros/service/CobrancaPlusService.java
@@ -47,7 +47,8 @@ public class CobrancaPlusService {
         }
     }
 
-    public void Remover(Long id){
+    public void Remover(Integer id){
+        if(!cobrancaAdicionalRepository.existsById(id)) throw new RecursoNaoEncontrado("Cobranca não encontrada");
         cobrancaAdicionalRepository.deleteById(id);
     }
